공지사항 등록 close버튼 삭제 + 공지사항 조회 추가
